modify #32 Form系のComponentを別ファイル化、さらにPureComponentに変更
